feat: refactor search strategy imports and restructure base module
